chore: remove source applications functionality
